VehicleVerse-backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
Pillow==10.1.0
pydot==1.4.2
pyparsing==3.1.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
